pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loomllm"
version = "1.0.0"
description = "The Iterative LLM Framework — Write → Review → Refine → Repeat. Automatically."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "LoomLLM Contributors"},
]
keywords = [
    "llm", "ai", "agent", "multi-agent", "collaboration",
    "openai", "deepseek", "gemini", "langchain-alternative",
    "iterative-refinement", "quality-gate", "cost-aware",
    "multi-provider", "auto-review", "closed-loop",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/chenxinghang-a/LoomLLM"
Documentation = "https://github.com/chenxinghang-a/LoomLLM#readme"
Repository = "https://github.com/chenxinghang-a/LoomLLM"
Issues = "https://github.com/chenxinghang-a/LoomLLM/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "build>=1.0",
    "twine>=5.0",
]

[project.scripts]
loomllm = "ai_staff_v4.__main__:main"

[tool.setuptools.packages.find]
include = ["ai_staff_v4*"]

[tool.setuptools.package-data]
ai_staff_v4 = ["experts/*.yaml", "py.typed"]
